Update index.html
set the icons and elements that are currently clored in primary colors to a purple-red color, add animations to the website as the user scrolls down, change the titlebar to say Portfolio -Scott Rafferty instead of just Scott Rafferty and make the titlebar elements align to each side of the screen. in mes competences add an extra bloc that says Tableau de competences with a button that opens (use the same built in pdf viewer as for the projects and certif tabs) a pdf called "tab_comp" stored in the same pdf folder as the other files. sort the "mon parcour" events in chronological order older events at the top. specify 1ere annéé in the BTS SIO option sisr event and make it appear chronologically before the microtel stage, add a BTS SIO option sisr 2 eme annéé event set in 2025-2026 chonologically set after the microtel stage. move the information contact box (the one containing phone number, mail etc) to the left of the message area (the one allowing users to send messages) and move the map to the right of that message area so that they are all on the same level. make the map point to lycée felix le dantec lannion instead of just centre ville lannion and remove the "mon lieu de residence" tag from the map ping icon